requirements.txt
alembic==1.0.11
aniso8601==7.0.0
attrs==19.1.0
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.14
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
PyJWT==1.7.1
pyrsistent==0.15.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.5
